refactor to es6 & clean up
