+++ index.d.ts
@@ -54,7 +54,6 @@ export interface Sift {
     <T extends SupportedTypes[]>(query: SiftQuery<T>): FilterFn;
     <T extends SupportedTypes>(query: SiftQuery<T>, target: T, rawSelector?: any): T;
     indexOf<T extends SupportedTypes>(query: SiftQuery<T>, target: T, rawSelector?: any): number;
-    use<K>(plugin: PluginFunction<K> | PluginDefinition<K>): void;
     compare<T, K>(a: T, b: K): 0 | -1 | 1;
 }
